Ignore non current spec levels
This is meant to test whether sticking to syntax definitions in "current" specs
would yield a result that is more closely aligned with the contents of MDN
data. There are a few places where that is indeed the case, but also a few
cases where the syntax in MDN data actually integrates changes made in not yet
current spec levels.